pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncpg==0.30.0",
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "pynacl>=1.5.0",
    "requests==2.31.0",
    "tlgbotfwk>=0.4.61",
]
